+++ src/main/java/med/voll/api/infra/http/controllers/inputs/doctor/DoctorSingUpInput.java
@@ -1,0 +1,18 @@
+package med.voll.api.infra.http.controllers.inputs.doctor;
+
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
+import med.voll.api.infra.http.controllers.inputs.SpecialtyInput;
+import med.voll.api.infra.http.controllers.inputs.address.AddressSignUpInput;
+
+public record DoctorSingUpInput(
+        @NotBlank String name,
+        @NotBlank @Email String email,
+        @NotBlank String phone,
+        @NotBlank @Pattern(regexp = "\\d{4,6}") String crm,
+        @NotNull SpecialtyInput specialty,
+        @NotNull @Valid AddressSignUpInput address) {
+}
