+++ src/main/java/med/voll/api/infra/http/controllers/inputs/doctor/DoctorUpdateInput.java
@@ -1,0 +1,11 @@
+package med.voll.api.infra.http.controllers.inputs.doctor;
+
+import jakarta.validation.Valid;
+import med.voll.api.infra.http.controllers.inputs.address.AddressUpdateInput;
+
+public record DoctorUpdateInput(
+                String id,
+                String name,
+                String phone,
+                @Valid AddressUpdateInput address) {
+}
