+++ src/main/java/med/voll/api/infra/http/controllers/output/commons/IdOutput.java
@@ -1,0 +1,4 @@
+package med.voll.api.infra.http.controllers.output.commons;
+
+public record IdOutput(String id) {
+}
