+++ src/main/java/med/voll/api/infra/http/controllers/output/commons/PaginatedOutput.java
@@ -1,4 +1,4 @@
-package med.voll.api.presentation.controllers.output.commons;
+package med.voll.api.infra.http.controllers.output.commons;
 
 public record PaginatedOutput<T>(int page, int limit, long total, T data) {
 }
