--- src/main/java/med/voll/api/presentation/controllers/inputs/address/AddressUpdateInput.java
@@ -1,14 +1,0 @@
-package med.voll.api.presentation.controllers.inputs.address;
-
-import jakarta.validation.constraints.Pattern;
-
-public record AddressUpdateInput(
-                String street,
-                String neighborhood,
-                @Pattern(regexp = "\\d{8}") String zip,
-                String city,
-                String state,
-                String number,
-                String complement) {
-
-}
