--- src/main/java/med/voll/api/presentation/controllers/inputs/doctor/DoctorSingUpInput.java
@@ -1,18 +1,0 @@
-package med.voll.api.presentation.controllers.inputs.doctor;
-
-import med.voll.api.presentation.controllers.inputs.address.AddressSignUpInput;
-import jakarta.validation.Valid;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Pattern;
-import med.voll.api.presentation.controllers.inputs.SpecialtyInput;
-
-public record DoctorSingUpInput(
-                @NotBlank String name,
-                @NotBlank @Email String email,
-                @NotBlank String phone,
-                @NotBlank @Pattern(regexp = "\\d{4,6}") String crm,
-                @NotNull SpecialtyInput specialty,
-                @NotNull @Valid AddressSignUpInput address) {
-}
