--- src/main/java/med/voll/api/presentation/controllers/output/commons/IdOutput.java
@@ -1,4 +1,0 @@
-package med.voll.api.presentation.controllers.output.commons;
-
-public record IdOutput(String id) {
-}
